+++ src/main/java/io/github/andreaspapadakis/banking/backoffice/accounts/dto/AccountUpdateRequest.java
@@ -1,7 +1,7 @@
 package io.github.andreaspapadakis.banking.backoffice.accounts.dto;
 
-import io.github.andreaspapadakis.banking.backoffice.accounts.validation.AllowedCurrency;
-import io.github.andreaspapadakis.banking.backoffice.accounts.validation.AtLeastOneFieldPresent;
+import io.github.andreaspapadakis.banking.backoffice.shared.validation.AllowedCurrency;
+import io.github.andreaspapadakis.banking.backoffice.shared.validation.AtLeastOneFieldPresent;
 import jakarta.validation.constraints.Pattern;
 
 @AtLeastOneFieldPresent
