+++ src/main/java/io/github/andreaspapadakis/banking/backoffice/accounts/validator/AtLeastOneAccountFieldPresentValidator.java
@@ -1,13 +1,12 @@
 package io.github.andreaspapadakis.banking.backoffice.accounts.validator;
 
 import io.github.andreaspapadakis.banking.backoffice.accounts.dto.AccountUpdateRequest;
-import io.github.andreaspapadakis.banking.backoffice.accounts.validation.AtLeastOneFieldPresent;
 import io.github.andreaspapadakis.banking.backoffice.shared.utils.StringUtils;
-import jakarta.validation.ConstraintValidator;
+import io.github.andreaspapadakis.banking.backoffice.shared.validator.AtLeastOneFieldPresentValidator;
 import jakarta.validation.ConstraintValidatorContext;
 
-public class AtLeastOneFieldPresentValidator implements
-    ConstraintValidator<AtLeastOneFieldPresent, AccountUpdateRequest> {
+public class AtLeastOneAccountFieldPresentValidator extends
+    AtLeastOneFieldPresentValidator<AccountUpdateRequest> {
   @Override
   public boolean isValid(AccountUpdateRequest accountUpdateRequest,
                          ConstraintValidatorContext context) {
