+++ src/main/java/io/github/andreaspapadakis/banking/backoffice/shared/validator/AllowedCurrencyValidator.java
@@ -1,6 +1,7 @@
-package io.github.andreaspapadakis.banking.backoffice.accounts.validator;
+package io.github.andreaspapadakis.banking.backoffice.shared.validator;
 
-import io.github.andreaspapadakis.banking.backoffice.accounts.validation.AllowedCurrency;
+import io.github.andreaspapadakis.banking.backoffice.shared.utils.StringUtils;
+import io.github.andreaspapadakis.banking.backoffice.shared.validation.AllowedCurrency;
 import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
 import java.util.Currency;
@@ -8,8 +9,8 @@ import java.util.Currency;
 public class AllowedCurrencyValidator implements ConstraintValidator<AllowedCurrency, String> {
   @Override
   public boolean isValid(String currency, ConstraintValidatorContext context) {
-    if (currency == null) {
-      return true; // @NotBlank handles null/empty
+    if (StringUtils.isNullOrEmpty(currency)) {
+      return true; // @NotBlank should handle null/empty
     }
 
     try {
